Pass "-machine" options through to armasm
Filter them out from the preprocessor invocation, and don't remove
them based on matching the "^-m" pattern.

As an alternative, we could also have gas-preprocessor.pl implicitly
add this option if the option "-arm64EC" is found, but requiring
the user to pass "-machine arm64ec" explicitly is more transparent.