Translate .inst into DCD/DCW, like .int/.short
For ELF targets, the assembler keeps track of which words are
actual instructions, and which are data - and the disassembler
shows them differently depending on this. For other object file
formats, there's no difference. Thus, for armasm for targeting
COFF, we can treat this just like any regular integer.

Produce either a 16 or 32 bit value depending on whether we
are assembling thumb or arm. (In the case of aarch64, it is always
32 bit.)